Trait pallet_committee::pallet::Config

pub trait Config: Config {
    type Origin: From<CommitteeOrigin<Self::AccountId, Self::BlockNumber>>;
    type Action: Parameter + Dispatchable<Origin = <Self as Config>::Origin, PostInfo = PostDispatchInfo> + From<Call<Self>> + GetDispatchInfo + TypeInfo;
    type ProposalNonce: Parameter + Member + One + Zero + Codec + Default + MaybeSerializeDeserialize + CheckedAdd + TypeInfo;
    type ProposalSubmissionPeriod: Get<Self::BlockNumber>;
    type VotingPeriod: Get<Self::BlockNumber>;
    type VotingPeriodRange: VotingPeriodRange<Self::BlockNumber>;
    type MinCouncilVotes: Get<usize>;
    type ProposalSubmissionOrigin: EnsureOrigin<<Self as Config>::Origin, Success = <Self as Config>::AccountId>;
    type ProposalExecutionOrigin: EnsureOrigin<<Self as Config>::Origin, Success = <Self as Config>::AccountId>;
    type ApprovedByCommitteeOrigin: EnsureOrigin<<Self as Config>::Origin>;
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Origin: From<CommitteeOrigin<Self::AccountId, Self::BlockNumber>>

The outer origin type.

type Action: Parameter + Dispatchable<Origin = <Self as Config>::Origin, PostInfo = PostDispatchInfo> + From<Call<Self>> + GetDispatchInfo + TypeInfo

The outer call dispatch type.

type ProposalNonce: Parameter + Member + One + Zero + Codec + Default + MaybeSerializeDeserialize + CheckedAdd + TypeInfo

A unique number assigned to each new instance of a proposal

type ProposalSubmissionPeriod: Get<Self::BlockNumber>

Duration (in blocks) of the proposal submission period

type VotingPeriod: Get<Self::BlockNumber>

Duration (in blocks) of the voting period

type VotingPeriodRange: VotingPeriodRange<Self::BlockNumber>

Range of the voting period

type MinCouncilVotes: Get<usize>

Minimum number of council members that must vote for a action to be passed

type ProposalSubmissionOrigin: EnsureOrigin<<Self as Config>::Origin, Success = <Self as Config>::AccountId>

Origin that is permitted to create proposals

type ProposalExecutionOrigin: EnsureOrigin<<Self as Config>::Origin, Success = <Self as Config>::AccountId>

Origin that is permitted to execute approved proposals

type ApprovedByCommitteeOrigin: EnsureOrigin<<Self as Config>::Origin>

Origin that is permitted to execute priviliged extrinsics

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type WeightInfo: WeightInfo

The weight for this pallet’s extrinsics.