Module pallet_committee::types

Structs

CommitteeMember

Assignment of a member type to an accountId

EnsureApprovedByCommittee

An implementation of EnsureOrigin approval of the committee

EnsureMember

Ensure committee member

MemberVote

A committee member together with their cast vote.

Proposal

This represents an instance of a proposal that can be voted on. It has been proposed and has an assigned nonce. This extra abstraction is required since it may be desirable construct multiple proposal instances out of a single proposal

VoteAggregate

Info for keeping track of a motion being voted on. Default is empty vectors for all votes

Enums

CommitteeOrigin

Origin for the committee pallet.

MemberType

Defines what sub-type a member belongs to. Council members are fixed in number and can vote on proposals Constituent members are unbounded in number but can only veto council proposals

ProposalStatus

VoteKind

Possible votes a member can cast

VoteRejectionReason